Stop using anonymous classes.
They make stack traces harder to understand.  With all the classes named
`Deferred$1', `Deferred$2' and whatnot, it's hard to keep track of which
one is which.

Change the code implementing join() slightly to avoid creating an
unnecessary array.

The new code is semantically equivalent to the old one.  As a bonus, the
uncompressed size of the byte code generated is 5% smaller on my laptop.

Change-Id: I4a9a743f6ab8213533544038bda5aaff91bc5268