Use a single array instead of 2 LinkedList for callbacks.
In retrospect, using LinkedList was a bad idea because of the insane
amount of memory overhead it has, especially in Java.  An array can
provide the same O(1) tail insertion (amortized), O(1) head removal,
while using far less RAM.

Heap dumps in OpenTSDB frequently show a lot of memory used just for
the LinkedList head and for each LinkedList$Entry.

On x86-64, with Sun's JVM 1.6.x, this yields the following savings:
  - First callback added costs an extra 32 bytes of RAM, instead of
    144 bytes with the LinkedList implementation.
  - Second callback added doesn't cost anything extra, whereas the
    LinkedList implementation adds 48 bytes of overhead per callback.
For a typical Deferred with 2 callbacks, this translates into 160 bytes
of RAM saved per instance.

This change also improves CPU cache efficiency since now all the
references to the callbacks are in the same cacheline (or contiguous
cachelines).  The end result is that a benchmark that creates 100k
Deferred, with an equal number of Deferred with 1, 2, 3, 4 and 5
callbacks, and calls them all back runs 14-15% faster on a dual Intel
E5520, 12-16% faster on dual Intel L5630.

Change-Id: I1c5e487947fe6e478940f16172bf90a1e4021298