Fix a race condition in Deferred.toString()
The `result' field was checked against `null' and then used.  However
this field can change at any time, so we must first copy it into a local
variable to make sure it doesn't become `null' at an incongruous moment,
which could cause a `NullPointerException'.  This issue is especially
likely to occur when a `TimeoutException' is raised roughly at the same
time as the `result' field is being updated by a callback.

Bug reported by Viral Bajaria.