Deferred should not be final
We use the Async Kudu Client (and I have used the Async HBase client before), and we have issues testing this in unit tests because the Deferred class is final. To mock it, we have to jump through hoops with power mock. I don't see a reason for this class to be final, so just removing the final keyword.